Public Transport Mobility Simulation
Design and implement a multi-agent system that represents the movements of passengers and public transport vehicles in a city. The city is described at a minimum by roads and bus stops. Passengers have an origin
and a destination in the city that they can reach by foot or by public transport. The public transport follow
predefined timetables. The objective is to simulate the movements of passengers and vehicles. In a second step,
we would like to introduce disturbances in the system. The disturbances can be a disconnection of a road or
the breakdown of a vehicle.